requirements.txt
requests
socket
ssl
hashlib
ipwhois
base64
mmh3
paramiko
logging
dns.resolver